--- data-index/data-index-service/src/main/java/org/kie/kogito/index/storage/Producer.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.storage;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Any;
-import javax.enterprise.inject.Instance;
-import javax.enterprise.inject.Produces;
-import javax.inject.Inject;
-
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-import org.kie.kogito.index.cache.CacheService;
-
-@ApplicationScoped
-public class Producer {
-
-    @ConfigProperty(name = "kogito.data-index.storage.type")
-    String storageType;
-
-    @Inject
-    @Any
-    Instance<CacheService> cacheServices;
-
-    @Produces
-    public CacheService cacheService() {
-        return cacheServices.select(new StorageImpl(storageType)).get();
-    }
-}
