--- data-index/data-index-service/src/main/java/org/kie/kogito/index/storage/StorageImpl.java
@@ -1,55 +1,0 @@
-/*
- * Copyright 2020 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.storage;
-
-import java.util.Objects;
-
-import javax.enterprise.util.AnnotationLiteral;
-
-public class StorageImpl extends AnnotationLiteral<Storage> implements Storage {
-
-    private final String type;
-
-    StorageImpl(final String type) {
-        this.type = type;
-    }
-
-    @Override
-    public String value() {
-        return type;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        if (!super.equals(o)) {
-            return false;
-        }
-        StorageImpl storage = (StorageImpl) o;
-        return Objects.equals(type, storage.type);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), type);
-    }
-}
