--- data-index/data-index-storage/data-index-storage-api/src/main/java/org/kie/kogito/index/cache/Cache.java
@@ -1,35 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates. 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.cache;
-
-import java.util.concurrent.ConcurrentMap;
-import java.util.function.Consumer;
-
-import org.kie.kogito.index.query.Query;
-
-public interface Cache<K, V> extends ConcurrentMap<K, V> {
-
-    void addObjectCreatedListener(Consumer<V> consumer);
-
-    void addObjectUpdatedListener(Consumer<V> consumer);
-
-    void addObjectRemovedListener(Consumer<K> consumer);
-    
-    Query<V> query();
-    
-    String getRootType();
-}
