--- data-index/data-index-storage/data-index-storage-api/src/main/java/org/kie/kogito/index/domain/AttributeDescriptor.java
@@ -1,75 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates. 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.domain;
-
-import java.util.Objects;
-
-public class AttributeDescriptor {
-
-    private String name;
-    private String typeName;
-
-    public AttributeDescriptor() {
-    }
-
-    public AttributeDescriptor(String name, String typeName) {
-        this.name = name;
-        this.typeName = typeName;
-    }
-
-    public String getTypeName() {
-        return typeName;
-    }
-
-    public void setTypeName(String typeName) {
-        this.typeName = typeName;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    @Override
-    public String toString() {
-        return "AttributeDescriptor{" +
-                "name='" + name + '\'' +
-                ", typeName='" + typeName + '\'' +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        AttributeDescriptor that = (AttributeDescriptor) o;
-        return Objects.equals(name, that.name) &&
-                Objects.equals(typeName, that.typeName);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, typeName);
-    }
-}
