--- data-index/data-index-storage/data-index-storage-api/src/main/java/org/kie/kogito/index/domain/DomainDescriptor.java
@@ -1,68 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates. 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.domain;
-
-import java.util.List;
-import java.util.Objects;
-
-public class DomainDescriptor {
-
-    private String typeName;
-    private List<AttributeDescriptor> attributes;
-
-    public List<AttributeDescriptor> getAttributes() {
-        return attributes;
-    }
-
-    public void setAttributes(List<AttributeDescriptor> attributes) {
-        this.attributes = attributes;
-    }
-
-    public String getTypeName() {
-        return typeName;
-    }
-    
-    public void setTypeName(String typeName) {
-        this.typeName = typeName;
-    }
-
-    @Override
-    public String toString() {
-        return "DomainDescriptor{" +
-                "typeName='" + typeName + '\'' +
-                ", attributes=" + attributes +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        DomainDescriptor that = (DomainDescriptor) o;
-        return Objects.equals(typeName, that.typeName) &&
-                Objects.equals(attributes, that.attributes);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(typeName, attributes);
-    }
-}
