--- data-index/data-index-storage/data-index-storage-api/src/main/java/org/kie/kogito/index/query/Query.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates. 
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.kie.kogito.index.query;
-
-import java.util.List;
-
-public interface Query<T> {
-
-    Query<T> limit(Integer limit);
-
-    Query<T> offset(Integer offset);
-
-    Query<T> filter(List<AttributeFilter> filters);
-
-    Query<T> sort(List<AttributeSort> sortBy);
-    
-    List<T> execute();
-
-}
